ppg_reconstruction_biometrics_verify/requirements-tf.txt
tensorflow
numpy
scipy
matplotlib
scikit-learn
tqdm
pyyaml
